Move custom filters into seperate project
