+++ src/Ensconce.NDjango.Custom/Filters/DefaultFilter.cs
@@ -1,6 +1,6 @@
 ﻿using Ensconce.NDjango.Core;
 
-namespace Ensconce.Update.NDjango.Custom.Filters
+namespace Ensconce.NDjango.Custom.Filters
 {
     [Interfaces.Name("default")]
     public class DefaultFilter : Interfaces.IFilter
@@ -12,7 +12,7 @@ namespace Ensconce.Update.NDjango.Custom.Filters
 
         public object PerformWithParam(object value, object parameter)
         {
-            return value is NDjangoWrapper.ErrorTemplate ? parameter : value;
+            return value is ErrorTemplate ? parameter : value;
         }
 
         public object DefaultValue
