+++ src/Ensconce.NDjango.Custom/Filters/EmptyFilter.cs
@@ -1,13 +1,13 @@
 ﻿using Ensconce.NDjango.Core;
 
-namespace Ensconce.Update.NDjango.Custom.Filters
+namespace Ensconce.NDjango.Custom.Filters
 {
     [Interfaces.Name("empty")]
     public class EmptyFilter : Interfaces.ISimpleFilter
     {
         public object Perform(object value)
         {
-            return value is NDjangoWrapper.ErrorTemplate || string.IsNullOrWhiteSpace(value.ToString());
+            return value is ErrorTemplate || string.IsNullOrWhiteSpace(value.ToString());
         }
     }
 }
