+++ src/Ensconce.NDjango.Custom/Filters/ExistsFilter.cs
@@ -1,13 +1,13 @@
 ﻿using Ensconce.NDjango.Core;
 
-namespace Ensconce.Update.NDjango.Custom.Filters
+namespace Ensconce.NDjango.Custom.Filters
 {
     [Interfaces.Name("exists")]
     public class ExistsFilter : Interfaces.ISimpleFilter
     {
         public object Perform(object value)
         {
-            return !(value is NDjangoWrapper.ErrorTemplate);
+            return !(value is ErrorTemplate);
         }
     }
 }
