+++ src/Ensconce.sln
@@ -47,13 +47,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ensconce.ReportingServices.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ensconce.Console", "Ensconce.Console\Ensconce.Console.csproj", "{560042FF-AF53-4925-9426-CA1D9AA4AD1E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ensconce.Cli", "Ensconce.Cli\Ensconce.Cli.csproj", "{5F6BB210-9BFC-465B-8348-FBC8CC78896F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ensconce.Cli", "Ensconce.Cli\Ensconce.Cli.csproj", "{5F6BB210-9BFC-465B-8348-FBC8CC78896F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ensconce.DotNetTool", "Ensconce.DotNetTool\Ensconce.DotNetTool.csproj", "{F3704257-E6FC-409E-A3F6-FCAE3D60336E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ensconce.DotNetTool", "Ensconce.DotNetTool\Ensconce.DotNetTool.csproj", "{F3704257-E6FC-409E-A3F6-FCAE3D60336E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ensconce.Cake", "Ensconce.Cake\Ensconce.Cake.csproj", "{62B30B97-A267-4832-B241-71BE5EE3BE22}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ensconce.Cake", "Ensconce.Cake\Ensconce.Cake.csproj", "{62B30B97-A267-4832-B241-71BE5EE3BE22}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ensconce.Helpers", "Ensconce.Helpers\Ensconce.Helpers.csproj", "{4E35B578-FFB8-4A9F-8E1D-DCCAEA15213F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ensconce.Helpers", "Ensconce.Helpers\Ensconce.Helpers.csproj", "{4E35B578-FFB8-4A9F-8E1D-DCCAEA15213F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ensconce.NDjango.Custom", "Ensconce.NDjango.Custom\Ensconce.NDjango.Custom.csproj", "{07E449DA-15D3-494F-A6CF-26C56B5620D3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -233,6 +235,18 @@ Global
 		{4E35B578-FFB8-4A9F-8E1D-DCCAEA15213F}.Release|x64.Build.0 = Release|Any CPU
 		{4E35B578-FFB8-4A9F-8E1D-DCCAEA15213F}.Release|x86.ActiveCfg = Release|Any CPU
 		{4E35B578-FFB8-4A9F-8E1D-DCCAEA15213F}.Release|x86.Build.0 = Release|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Debug|x64.Build.0 = Debug|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Debug|x86.Build.0 = Debug|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Release|x64.ActiveCfg = Release|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Release|x64.Build.0 = Release|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Release|x86.ActiveCfg = Release|Any CPU
+		{07E449DA-15D3-494F-A6CF-26C56B5620D3}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
